feat(index.html, index.js, preload.js): add widescreen mode toggle functionality to the app
This commit adds the ability to toggle widescreen mode in the app. When the user presses the 'f' key, the app toggles widescreen mode on or off. The widescreen mode removes the max-width from each chat message and adds padding to the left and right of the chat messages so that they don't take up the whole screen. The widescreen mode is toggled by injecting a CSS style element into the webview. The commit also includes some minor refactoring and code cleanup.